# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SCORM Engine API
  version: 0.1.0
  summary: Modern SCORM runtime + tracking engine for TypeScript applications.
  description: |
    The SCORM Engine API is a **vendor-neutral REST contract** for SCORM 1.2 and
    SCORM 2004 course ingestion, runtime tracking, and analytics. It is the
    primary contract for our client SDK (`scorm-engine-client`), but the API is
    designed to be consumed by **any HTTP client**.

    The engine ships static-asset routes (`/play/:attemptId`, `/runtime.js`) that
    bridge browser-side SCOs to the API. The runtime bridge is a **clean-room,
    spec-compliant implementation of the SCORM JavaScript API** — it has no
    third-party SCORM dependencies. Because it follows the spec, any SCORM-compliant
    course works regardless of the authoring tool or API wrapper it was built with.

    ## Versioning
    The API is versioned via the URL prefix (`/api/v1`). Breaking changes ship under
    a new prefix. Additive changes (new optional fields, new endpoints) are non-breaking.

    ## Authentication
    Two auth modes:
    - **API key** (`X-API-Key` header): service-to-service calls from an LMS backend.
    - **Attempt JWT** (`Authorization: Bearer <jwt>`): short-lived token scoped to
      a single attempt, used by the in-browser player. Issued by `/auth/attempt-token`.

    ## Tenancy
    Every resource is scoped to a tenant resolved from the API key. Attempt JWTs
...
  contact:
    name: SCORM Engine
...
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /courses/{courseId}/attempts:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    post:
      tags: [Attempts]
      summary: Start or resume an attempt
      description: |
        Starts a new attempt for a learner against the course, or resumes the
        learner's existing in-progress attempt for the same SCO. A fresh attempt
        launches `ab-initio`; a resumed one launches `resume` and returns its last
        committed CMI snapshot so the player can rehydrate the SCO.

        Pass `restart: true` to force a brand-new attempt even when a resumable
        one exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [learnerId]
              properties:
                learnerId:
                  type: string
                  description: |
                    Opaque identifier from the host LMS. The engine never stores PII —
                    this is the only learner identifier it knows.
                learnerName:
                  type: string
                  description: Optional display name passed to the SCO as `student_name`.
                scoId:
                  type: string
                  description: Target SCO; defaults to the course's primary SCO.
                mode:
                  type: string
                  enum: [normal, browse, review]
                  default: normal
                restart:
                  type: boolean
                  default: false
                  description: Force a new attempt instead of resuming an open one.
      responses:
        '201':
          description: New attempt created (ab-initio).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StartAttemptResponse' }
        '200':
          description: Existing in-progress attempt resumed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StartAttemptResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /attempts/{attemptId}:
    parameters:
      - $ref: '#/components/parameters/AttemptId'
    get:
      tags: [Attempts]
      summary: Get an attempt
      description: Returns the attempt's rolled-up state plus its objective and interaction rows.
      responses:
        '200':
          description: Attempt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Attempt' }
        '404': { $ref: '#/components/responses/NotFound' }
  /attempts/{attemptId}/runtime:
    parameters:
      - $ref: '#/components/parameters/AttemptId'
    get:
      tags: [Runtime]
      summary: Get the runtime state for (re)hydrating a SCO
      description: |
        Returns the CMI snapshot, SCORM version, and entry mode a player needs to
        initialize the SCO. `entry` is `resume` once the attempt has any committed
        CMI, otherwise `ab-initio`. This is the read side of the runtime loop; use
        the start-attempt response for the initial launch and this endpoint to
        re-fetch state (e.g. after a reload) without creating a new attempt.

        Auth is dual: a tenant API key, or an attempt token scoped to this attempt.
      security:
        - ApiKeyAuth: []
        - AttemptJWT: []
      responses:
        '200':
          description: Runtime state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuntimeState' }
        '403':
          description: Token is not valid for this attempt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────────────────────────────────────────────────────────────────
  # Runtime
  # ─────────────────────────────────────────────────────────────────────────
  # Design note: a SCORM SCO can issue dozens of SetValue calls per minute.
  # Rather than one round-trip per call, the in-browser player holds CMI state
  # in memory and flushes changes via a single batched /commit. Initialize is
  # implicit in the start-attempt response (which carries the resume snapshot),
  # and Terminate is folded into /commit via the `terminate` flag.
  /attempts/{attemptId}/commit:
    parameters:
      - $ref: '#/components/parameters/AttemptId'
    post:
      tags: [Runtime]
      summary: Commit a batch of CMI value changes
      description: |
        Replays a batch of `SetValue` writes through an authoritative server-side
        runtime session seeded from the attempt's last snapshot, then persists the
        result. The player should call this:
        - On `LMSCommit` / `Commit` from the SCO.
        - On a debounced interval (e.g. every 10s) to reduce data-loss risk.
        - On `beforeunload` (with `keepalive: true`).

        Set `terminate: true` to close the session out (`LMSFinish` / `Terminate`),
        performing a final commit and finalizing the attempt.

        Keys use SCORM-version-native naming:
        - SCORM 1.2: `cmi.core.lesson_status`, `cmi.suspend_data`, ...
        - SCORM 2004: `cmi.completion_status`, `cmi.success_status`, ...

        **Validation policy** (server-configured via `SCORM_VALIDATION_MODE`):
        - `lenient` (default): mirrors certified commercial LMSes — writes that
          fail validation are still persisted and reported in `warnings`. SCORM
          content in the wild routinely violates the spec, and rejecting commits
          would break otherwise-functional courses. A single bad write never
          aborts the batch, so `ok` stays `true`.
        - `strict`: writes that fail validation are dropped and reported in
          `errors` (`ok` is `false`), for conformance-testing content. Valid
          writes in the same batch still persist.

        Either way the commit as a whole succeeds; inspect `errors`/`warnings` for
        per-element detail, not for retry control flow.

        Auth is dual: a tenant API key (server-side), or an attempt token scoped
        to this attempt (browser player).
      security:
        - ApiKeyAuth: []
        - AttemptJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitRequest'
      responses:
        '200':
          description: Commit accepted (possibly with warnings or per-element errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitResponse'
        '403':
          description: Token is not valid for this attempt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Attempt already finished.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ─────────────────────────────────────────────────────────────────────────
  # Analytics
  # ─────────────────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Long-lived tenant API key. Used for server-to-server calls.
    AttemptJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT scoped to a single attempt. Issued via /auth/attempt-token.
        Used by the in-browser player.

  # ───────────────────────────────────────────────────────────────────────
  # Reusable parameters
  # ───────────────────────────────────────────────────────────────────────
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema: { type: string, format: uuid }
    AttemptId:
      name: attemptId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Request was malformed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code (e.g. `course_not_found`).
        message: { type: string }
        details:
          type: object
          additionalProperties: true
        requestId: { type: string }
...
    ScormVersion:
      type: string
      enum: [SCORM_1_2, SCORM_2004_2, SCORM_2004_3, SCORM_2004_4]

    # ── Course ───────────────────────────────────────────────────────────
...
    Attempt:
      type: object
      required: [id, courseId, learnerId, status, attemptNumber, startedAt]
      properties:
        id: { type: string, format: uuid }
        courseId: { type: string, format: uuid }
        learnerId: { type: string }
        learnerName: { type: string, nullable: true }
        attemptNumber:
          type: integer
          description: 1-indexed sequence of attempts by this learner on this course.
        status:
          type: string
          enum: [not_started, in_progress, completed, passed, failed, terminated]
        completionStatus:
          type: string
          enum: [unknown, completed, incomplete, not_attempted]
          description: SCORM 2004 completion status; mapped from `cmi.completion_status`.
        successStatus:
          type: string
          enum: [unknown, passed, failed]
          description: SCORM 2004 success status; mapped from `cmi.success_status`.
        score:
          $ref: '#/components/schemas/Score'
        progressMeasure:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        totalTimeSeconds:
          type: integer
          description: Total accumulated session time across all sessions in this attempt.
        location:
          type: string
          nullable: true
          description: Bookmark (`cmi.core.lesson_location` / `cmi.location`).
        startedAt: { type: string, format: date-time }
        lastAccessedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
    Score:
      type: object
      properties:
        raw: { type: number, nullable: true }
        min: { type: number, nullable: true }
        max: { type: number, nullable: true }
        scaled:
          type: number
          nullable: true
          minimum: -1
          maximum: 1
          description: SCORM 2004 normalized score (-1..1).

    # ── Runtime ──────────────────────────────────────────────────────────
    RuntimeState:
      type: object
      required: [attemptId, version, status, cmi, entry, learner]
      properties:
        attemptId: { type: string, format: uuid }
        version: { $ref: '#/components/schemas/ScormVersion' }
        status:
          type: string
          enum: [IN_PROGRESS, COMPLETED, TERMINATED, EXPIRED]
        cmi:
          type: object
          additionalProperties: true
          description: |
            Full CMI data model snapshot, keyed by SCORM-native paths. Empty object
            for a new attempt; for a resume, contains every previously-committed key.
          example:
            cmi.core.lesson_status: incomplete
            cmi.core.lesson_location: page_4
            cmi.suspend_data: '{"chapter":2,"answers":[1,3,2]}'
            cmi.core.score.raw: 42
        entry:
          type: string
          enum: [ab-initio, resume]
          description: SCORM `cmi.core.entry` / `cmi.entry`.
        learner:
          type: object
          required: [id]
          properties:
            id: { type: string }
            name: { type: string, nullable: true }
    StartAttemptResponse:
      type: object
      required: [attemptId, resumed, version, cmi]
      properties:
        attemptId: { type: string, format: uuid }
        resumed:
          type: boolean
          description: True when an existing in-progress attempt was resumed.
        version: { $ref: '#/components/schemas/ScormVersion' }
        entry:
          type: string
          enum: [ab-initio, resume, '']
          description: SCORM `cmi.core.entry` / `cmi.entry` the SCO should launch with.
        launch:
          type: object
          description: Read-only launch context the player injects before Initialize.
          properties:
            learnerId: { type: string }
            learnerName: { type: string }
            mode: { type: string, enum: [normal, browse, review] }
            totalTime:
              type: string
              description: Accumulated time from prior attempts, in the edition's timespan format.
            masteryScore: { type: number, nullable: true }
        cmi:
          type: object
          additionalProperties: true
          description: |
            Previously-committed CMI snapshot to rehydrate the SCO on resume. Empty
            object for a new (ab-initio) attempt.
          example:
            cmi.location: page_4
            cmi.suspend_data: '{"chapter":2}'
    CommitRequest:
      type: object
      required: [values]
      properties:
        values:
          oneOf:
            - type: object
              additionalProperties: { type: string }
            - type: array
              items:
                type: array
                minItems: 2
                maxItems: 2
                items: { type: string }
          description: |
            CMI writes to replay, as an object of key → value pairs or an ordered
            array of `[key, value]` pairs (use the array form when write order
            matters). Keys must match the SCORM version of the course.
          example:
            cmi.completion_status: completed
            cmi.score.scaled: '0.87'
            cmi.session_time: PT12M30S
        terminate:
          type: boolean
          default: false
          description: |
            When true, Terminate the session (final commit) and finalize the
            attempt instead of leaving it open for further commits.
    CommitResponse:
      type: object
      required: [ok, terminated, errors, warnings, summary]
      properties:
        ok:
          type: boolean
          description: False only in strict mode when one or more writes were rejected.
        terminated:
          type: boolean
          description: True when this commit finalized the attempt.
        errors:
          type: array
          description: |
            Writes rejected in strict mode (not persisted). Empty in lenient mode.
          items: { $ref: '#/components/schemas/CommitIssue' }
        warnings:
          type: array
          description: |
            Nonconformant writes accepted in lenient mode (persisted verbatim).
            Empty in strict mode. Use for observability, not retry logic.
          items: { $ref: '#/components/schemas/CommitIssue' }
        summary: { $ref: '#/components/schemas/RuntimeSummary' }
    CommitIssue:
      type: object
      required: [element, code, message]
      properties:
        element:
          type: string
          description: The CMI element the write targeted.
        code:
          type: integer
          description: SCORM API error code (edition-specific numeric code).
        message: { type: string }
    RuntimeSummary:
      type: object
      description: Edition-neutral rollup of the outcome-bearing CMI elements.
      properties:
        completionStatus:
          type: string
          enum: [completed, incomplete, 'not attempted', unknown]
        successStatus:
          type: string
          enum: [passed, failed, unknown]
        scoreRaw: { type: number, nullable: true }
        scoreMin: { type: number, nullable: true }
        scoreMax: { type: number, nullable: true }
        scoreScaled: { type: number, nullable: true }
        progressMeasure: { type: number, nullable: true }
        lessonLocation: { type: string, nullable: true }
        suspendData: { type: string, nullable: true }
        sessionTime: { type: string, nullable: true }
        exit: { type: string, nullable: true }

    # ── Analytics ────────────────────────────────────────────────────────
...